pyproject.toml
[project]
name = "news-agent"
version = "0.2.0"
description = "Kiara Health news pipeline: RSS ingestion, scoring, summarisation, and content generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # feed-ingestor deps (now part of this package)
    "feedparser>=6.0",
    "requests>=2.31",
    "tenacity>=8.2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    # news-agent deps
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
feed-ingestor = "ingestor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["ingestor"]
omit = ["tests/*"]

[tool.uv]
